% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/angleDist.R
\name{angleDist}
\alias{angleDist}
\title{Computes distance and angle from marker columns}
\usage{
angleDist(data, colNames, colNameSubj, rndDig = 2, verbose = FALSE)
}
\arguments{
\item{data}{Data frame containig the data for a subset, defined in colNames, 
for which the angle and the distance should be computed for.}

\item{colNames}{Character vector naming the two columns with the x/y-Values
for which the angle and the distance should be computed.}

\item{colNameSubj}{Character vector with a single value containig the name 
of the subject column of the data data-frame.}

\item{rndDig}{Number of round digits. Default is 2.}

\item{verbose}{If TRUE, the function prints verbose output. Otherwise not. 
Default is FALSE.}
}
\value{
Numeric matrix with median distance and angle.
}
\description{
Computes the distance and the angle for a (mean, median) x/y marker 
point. The distance and the angle is computed from x = 0, y = 0 to the 
x/y - values given. Therefore, to get correct values for distance and angle
use the function \code{\link{centerCond}} before, to set the x/y-start 
values for each condition to x = 0, y = 0.
}
\examples{
colNames <- c("A7_x",  "A7_y",  "A8_x",  "A8_y",  
             "BL2_x", "BL2_y", "BL4_x", "BL4_y",  
             "BL5_x", "BL5_y", "BL7_x", "BL7_y",        
             "BR2_x", "BR2_y", "BR4_x", "BR4_y",  
             "BR5_x", "BR5_y", "BR7_x", "BR7_y",  
             "CL4_x", "CL4_y", "CL7_x", "CL7_y",        
             "CR4_x", "CR4_y", "CR7_x", "CR7_y",  
             "DL2_x", "DL2_y", "DR2_x", "DR2_y")

# Data preparation
data_Subj_happy <- subset(dataSmm, subset = dataSmm["Stimulustype"] == "posed_happy",
                          select = c("subject",colNames))
angleDist(data_Subj_happy, colNames = c("CL7_x", "CL7_y"),
          colNameSubj = "subject", rndDig = 3)

}
\author{
Axel Zinkernagel \email{zinkernagel@uni-landau.de}
}
