% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_Function_AngleDistance.R
\name{angleDistance}
\alias{angleDistance}
\title{Computes distance and angle from marker columns}
\usage{
angleDistance(x, y, verbose = FALSE)
}
\arguments{
\item{x}{Numeric vector of x values of x marker movement, beginnig at 0}

\item{y}{Numeric vector of y values of y marker movement, beginnig at 0}

\item{verbose}{If TRUE, the function prints verbose output. Otherwise not. 
Default is FALSE.}
}
\value{
Numeric vector with median distance and angle.
}
\description{
Computes the distance and the angle for a (mean, median) x/y marker 
point. The distance and the angle is computed from x = 0, y = 0 to the 
x/y - values given. Therefore, to get correct values for distance and angle
use the function CenderCond() before, to set the x/y-start values for each 
condition to x = 0, y = 0.
}
\examples{
\dontrun{angleDistance(x,y)}

}
\author{
Axel Zinkernagel \email{zinkernagel@uni-landau.de}
}

