\name{blenderFace-package}
\alias{blenderFace-package}
\alias{blenderFace}
\docType{package}
\title{
\packageTitle{blenderFace}
}
\description{
\packageDescription{blenderFace}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{blenderFace}
\packageIndices{blenderFace}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{blenderFace}

Maintainer: \packageMaintainer{blenderFace}
}
\references{
~~ Literature or other references for background information ~~
}
%~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
%~~ directory ~~
\keyword{ package }
%\seealso{
%~~ Optional links to other man pages, e.g. ~~
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
%}
%\examples{
%~~ simple examples of the most important functions ~~
%}
