% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_Function_CenterCond.R
\name{CenterCond}
\alias{CenterCond}
\title{Center marker coordinates}
\usage{
CenterCond(data, colNames, colNameSubj, colNameFrames, colNameCond,
  verbose = FALSE)
}
\arguments{
\item{data}{Data frame containig the columns which should be centered.
Remaining colums will be returned untouched.}

\item{colNames}{Character vector with names of the columns which should be
centered.}

\item{colNameSubj}{Character vector with a single value containig the name of
the subject column. It is assumed, that each experimental condition
occurres only once per subject}

\item{colNameFrames}{Character vector with a single value containig the name
of the frames column.}

\item{colNameCond}{Character vector with a single value containing the
experimental conditions column name of the data frame.}

\item{verbose}{If TRUE, the function prints verbose output. Otherwise not.}
}
\value{
Data frame with centered columns.
}
\description{
Sets the x/y/z-values of markers in colNames defined variable columns to 0 at
the start of every condition defined in variable cond. This is done for every
subject marked in column colNameSubj.
}
\examples{
\dontrun{CenterCond()}

#@include doParallel

}
\author{
Axel Zinkernagel \email{zinkernagel@uni-landau.de}
}

