\name{dataSmm}
\alias{dataSmm}
\docType{data}
\title{
R-data file of data scaled to millimeter of the two sample subjects.
}
\description{
R-data file of data scaled to millimeter of the two sample subjects concatenated by the function \code{\link{concatBlenderFiles}}. Scaling has been performed with function the \code{\link{bu2mm}} for the `rawdata` dataset.
}
\usage{data("dataSmm")}
\format{
  A data frame with 3406 observations on the following 51 variables.
  \describe{
    \item{\code{BL2_x}}{x-axis movement in mm for Marker at BL2}
    \item{\code{BL2_y}}{y-axis movement in mm for Marker at BL2}
    \item{\code{BL2_z}}{z-axis movement in mm for Marker at BL2}
    \item{\code{BR2_x}}{x-axis movement in mm for Marker at BR2}
    \item{\code{BR2_y}}{y-axis movement in mm for Marker at BR2}
    \item{\code{BR2_z}}{z-axis movement in mm for Marker at BR2}
    \item{\code{DL2_x}}{x-axis movement in mm for Marker at DL2}
    \item{\code{DL2_y}}{y-axis movement in mm for Marker at DL2}
    \item{\code{DL2_z}}{z-axis movement in mm for Marker at DL2}
    \item{\code{DR2_x}}{x-axis movement in mm for Marker at DR2}
    \item{\code{DR2_y}}{y-axis movement in mm for Marker at DR2}
    \item{\code{DR2_z}}{z-axis movement in mm for Marker at DR2}
    \item{\code{CL4_x}}{x-axis movement in mm for Marker at CL4}
    \item{\code{CL4_y}}{y-axis movement in mm for Marker at CL4}
    \item{\code{CL4_z}}{z-axis movement in mm for Marker at CL4}
    \item{\code{CR4_x}}{x-axis movement in mm for Marker at CR4}
    \item{\code{CR4_y}}{y-axis movement in mm for Marker at CR4}
    \item{\code{CR4_z}}{z-axis movement in mm for Marker at CR4}
    \item{\code{A7_x}}{x-axis movement in mm for Marker at A7}
    \item{\code{A7_y}}{y-axis movement in mm for Marker at A7}
    \item{\code{A7_z}}{z-axis movement in mm for Marker at A7}
    \item{\code{BL4_x}}{x-axis movement in mm for Marker at BL4}
    \item{\code{BL4_y}}{y-axis movement in mm for Marker at BL4}
    \item{\code{BL4_z}}{z-axis movement in mm for Marker at BL4}
    \item{\code{BR4_x}}{x-axis movement in mm for Marker at BR4}
    \item{\code{BR4_y}}{y-axis movement in mm for Marker at BR4}
    \item{\code{BR4_z}}{z-axis movement in mm for Marker at BR4}
    \item{\code{BL7_x}}{x-axis movement in mm for Marker at BL7}
    \item{\code{BL7_y}}{y-axis movement in mm for Marker at BL7}
    \item{\code{BL7_z}}{z-axis movement in mm for Marker at BL7}
    \item{\code{BR7_x}}{x-axis movement in mm for Marker at BR7}
    \item{\code{BR7_y}}{y-axis movement in mm for Marker at BR7}
    \item{\code{BR7_z}}{z-axis movement in mm for Marker at BR7}
    \item{\code{BL5_x}}{x-axis movement in mm for Marker at BL5}
    \item{\code{BL5_y}}{y-axis movement in mm for Marker at BL5}
    \item{\code{BL5_z}}{z-axis movement in mm for Marker at BL5}
    \item{\code{BR5_x}}{x-axis movement in mm for Marker at BR5}
    \item{\code{BR5_y}}{y-axis movement in mm for Marker at BR5}
    \item{\code{BR5_z}}{z-axis movement in mm for Marker at BR5}
    \item{\code{CL7_x}}{x-axis movement in mm for Marker at CL7}
    \item{\code{CL7_y}}{y-axis movement in mm for Marker at CL7}
    \item{\code{CL7_z}}{z-axis movement in mm for Marker at CL7}
    \item{\code{CR7_x}}{x-axis movement in mm for Marker at CR7}
    \item{\code{CR7_y}}{y-axis movement in mm for Marker at CR7}
    \item{\code{CR7_z}}{z-axis movement in mm for Marker at CR7}
    \item{\code{A8_x}}{x-axis movement in mm for Marker at A8}
    \item{\code{A8_y}}{y-axis movement in mm for Marker at A8}
    \item{\code{A8_z}}{z-axis movement in mm for Marker at A8}
    \item{\code{Frame}}{Frame number of the video clip per participant}
    \item{\code{Stimulustype}}{Stimulus presented to the participant}
    \item{\code{subject}}{Subject number}
  }
}
\examples{
data(dataSmm)
}
\keyword{datasets}
