% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/face2stdFace.R
\name{face2stdFace}
\alias{face2stdFace}
\title{Scale head proportions}
\usage{
face2stdFace(data, colNames, colNameSubj, pupilDist, leftPMDist,
  mouthcornerDist = NA, rightPMDist = NA, rwMeasure = 8,
  verbose = FALSE)
}
\arguments{
\item{data}{Data frame containig the columns which should be rescaled.
Remaining colums will be returned untouched (e.g., z-values because we have
no information to rescale them).}

\item{colNames}{Character vector with names of the columns which should be
rescaled. To differ between x-, and y-axis scaling, the variable names containing horizontal movement (left-right) should end with '_x',
whereas variable names containing vertical movement (up-down) should end with '_y'.}

\item{colNameSubj}{Character vector with a single value containig the name 
of the subject column of the data data-frame.}

\item{pupilDist}{Numeric vector of measurement of the distance from left to right pupil for each subject,
measured in blender units.}

\item{leftPMDist}{Numeric vector of measuremet of the distance from left pupil to left
mouthcorner for each subject, measured in blender units.}

\item{mouthcornerDist}{Numeric vector of measurement of the distance from left to right
mouthcorner for each subject measured in blender units.}

\item{rightPMDist}{Numeric vector of measurement of the distance from right pupil to right
mouthcorner for each subject, measured in blender units.}

\item{verbose}{If TRUE, the function prints verbose output. Otherwise not.}
}
\value{
Data frame with scaled columns.
}
\description{
The heads of the participants are of different sizes. This function rescales
the x/y values according to a standardized face. The standardized face is of 
height 1 and width 1, while the pupil-pupil distance is defined as 1/3 and the 
left-pupil -- left-mouth-corner distance is defined as 1/3. Accordongly, for resizing
the x-axis, the distance between the pupils of the subject is used. For resizing the y-axis,
the distance between the left pupil and the left mouthcorner is used. 
Optionally, the distance between the mouthcorners for the x-axis, and the distance between
the right pupil and the right mouthcorner for the y-axis may additionally be provided.
If provided, a mean of both measures is computed. Be aware, if you previously rescaled 
the data into millimeter by the function \code{\link{bu2mm}}, you also have to
rescale the scaling parameters e.g., the pupil-pupil distance into millimeter first
and use the millimeter values as scaling parameters!
}
\examples{
\dontrun{face2stdFace()}

}
\author{
Axel Zinkernagel, Rainer Alexandrowicz
  \email{zinkernagel@uni-landau.de, rainer.alexandrowicz@aau.at}
}

