% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmParameters.R
\name{mmParameters}
\alias{mmParameters}
\title{Compute the marker movement parameters speed, onset-speed, intensity (amplitude), 
irregularites (onset, apex, offset-phases), symmetry)}
\usage{
mmParameters(
  data,
  colNames,
  colNameSubj,
  colNameFrames,
  mmCutoff = NA,
  verbose = FALSE
)
}
\arguments{
\item{data}{Dataframe containing the data with the columns subject, Frame and
marker names labeled according to the blenderFace-method (e.g., CL7_x, CL7_y, 
CR7_x, CR7_y; to compute symmetry values). If the dataframe privides values
on the z-axis (e.g., CL7_z), the eukledian distances (3D) are computed instead 
of the pythagorean distances (2D) to determine marker movement parameters.}

\item{colNames}{Character vector with marker names of the columns for which
the parameters should be computed. A touple or triplet of axes per marker is
needed (e.g., CL7_x, CL7_y or CL7_x, CL7_y, CL7_z) to compute parameters.
Mirrord markers (e.g., CL7_x, CL7_y and CR7_x, CR7_y) are needed to compute 
symmetry values.}

\item{colNameSubj}{Character vector with a single value containig the name 
of the subject column of the data data-frame.}

\item{colNameFrames}{Character vector containing the column name of the 
frames column of the data data-frame}

\item{mmCutoff}{Cutoff for defining a marker movement: It is only regarded to be a movement, if the 
distance from t to t+1 is larger than the value of mmCutoff. If NA (default) mmCutoff is computed
by the mean + 2SDs of the first 10 frames. However, it can be set to any value, e.g.,
to .5, which means a distance of .5 mm if data was scaled to mm via the function \code{\link{bu2mm}}.
If movement is smaller than mmCutoff, it is regarded to be an apex-phase or no movement.}

\item{verbose}{If TRUE, the function prints verbose output. Otherwise not. 
Default is FALSE.}
}
\value{
returns dataframe containing the marker movement parameters 
(speed, onst-speed, intensity, irregularities, symmetry) per subject.
}
\description{
Output Parameters:
}
\details{
Speed (sp): Mean distance from frame t to t+1

Onset-Speed (onSp): Mean distance from frame t to t+1 only for movement onset
phases (see irregularities)

Intensity/Amplitude (int): Mean of the 20 maximal distances from the first frame of 
data (= first frame is seen as baseline). Probably use \code{\link{centerCond}} 
in advance.

Irregularities (irr): A movement has to be larger than measurement noise and is 
defined to be larger than the baseline of .5 (which is .5 mm, if function 
\code{\link{bu2mm}} is used beforehand). An onset-phase occurs, if the 
distance between t+1 and t+2 is larger than the distance between t and t+1 (previous frame).
An apex-phase is defined by the distance between t and t+1 stays whithin
the baseline, independent of the direction of movement. An offset-phase occurs, if the distance between t+1 and t+2 
is smaller than the distance between t and t+1 (previous frame). Every change in phases is counted. Because the
analysed episodes have different numbers of frames per participant, the 
relative number of onset, apex, and offset-phases is computed.

Symmetry (sym): Mean correlation (aggregated over axes x,y, and z) of 
symmetrical markers. Symmetrical markers are determined by their label, e.g.,
if there is a marker label CL7 (for left mouth corner), it is tests if there 
is also a marker labeled CR7 (for right mouth corner). If this is true, the 
markers CL7 and CR7 are regarded to be symmetrical.
}
\examples{
markers <- c("CL7_x", "CL7_y", "CL7_z", "CR7_x", "CR7_y", "CR7_z")
myParams <- mmParameters(dataSmm, colNames = markers,
 colNameSubj = "subject",
  colNameFrames = "Frame",
  verbose = TRUE)
myParams

}
\author{
Axel Zinkernagel \email{zinkernagel@uni-wuppertal.de}
}
