% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotIndmm.R
\name{plotIndmm}
\alias{plotIndmm}
\title{Plots median movement of a marker per subject}
\usage{
plotIndmm(data, colNames, colNameSubj, title = "", xlim = c(-10, 10),
  ylim = c(-10, 10), verbose = FALSE)
}
\arguments{
\item{data}{Data frame containig the data to be plotted}

\item{colNames}{Character vector naming the two columns with the x/y-Values
to be plotted}

\item{colNameSubj}{Character vector with a single value containig the name 
of the subject column of the data data-frame.}

\item{title}{Character string containing the title. Default is no title.}

\item{xlim}{Numeric vector of two elements defining the x-axis range of the 
plot. Default is c(-10, 10).}

\item{ylim}{Numeric vector of two elements defining the y-axis range of the 
plot. Default is c(-10, 10).}

\item{verbose}{print verbose output}
}
\value{
Plot of median movement of marker with quartile ellipse around per
  condition
}
\description{
Plots median movement of a marker per subject and labels it with subject 
number. This function may be used to find individual outliers
}
\examples{
\dontrun{plotMmpCond()}
   
}
\author{
Axel Zinkernagel \email{zinkernagel@uni-landau.de}
}

