% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_Function_PlotMarkerCond.R
\name{plotMarkerCond}
\alias{plotMarkerCond}
\title{Plot marker and quartile ellipse per condition}
\usage{
plotMarkerCond(data, colNames, colNameCond, title = "", xlim = c(-10, 10),
  ylim = c(-10, 10), verbose = FALSE)
}
\arguments{
\item{data}{data frame containig the data to be plotted}

\item{colNames}{character vector naming the two columns with the x/y-Values
to be plotted}

\item{colNameCond}{character vector naming the column containig the
conditions}

\item{title}{Default is no title.}

\item{xlim}{Numeric vector of two elements defining the x-axis range of the 
plot. Default is c(-10, 10).}

\item{ylim}{Numeric vector of two elements defining the y-axis range of the 
plot. Default is c(-10, 10).}

\item{verbose}{print verbose output}
}
\value{
Plot of median movement of marker with quartile ellipse around per
  condition
}
\description{
Plots the median and quartile ellipse per experimental condition
}
\examples{
\dontrun{plotXYMarkerMovementPerFrame()}
   
}
\author{
Axel Zinkernagel \email{zinkernagel@uni-landau.de}
  
  #@include plotrix
}

