% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_Function_PlotXHead.R
\name{plotMarkerXHead}
\alias{plotMarkerXHead}
\title{Plot markers on symbolic head}
\usage{
plotMarkerXHead(data, dataPos, scaleFactor, title = "Square Head Baseplot",
  overplot = FALSE, color = "black", alpha = 0.5)
}
\arguments{
\item{data}{Data-frame with a row sequence of *standardized* x/y
coordinates for each marker. Please use the functions 'ScaleBlenderData' 
and 'ScaleFaceData' before plotting the data. Names of the data columns 
must and in '_x' for the x-axis and in '_y' for the y-axis.}

\item{dataPos}{numeric vector with x/y starting coordinates each marker
should be plotted on the face. This vector can also be generated by using 
the offsetTable generated by the function CenterCond(). Due to the tracking
procedure in blender, the center of the coordinate system is the nose tip 
at x = 0, y = 0. Scaling of the data to fit the plot is done automatically 
based on max./min. values of data columns.}

\item{scaleFactor}{Factor which is used to rescale data for x-head. Should be
a positive integer (e.g., 2000)}

\item{title}{Default is no title.}

\item{overplot}{If TRUE, only the marker plots, but no base graph is plotted.
Default is FALSE.}

\item{color}{character vector of lenght one to define the color for plotting
the Marker. Default is 'black'.}

\item{alpha}{alpha value / density for marker plots. 0 represents full 
transparency, whereas 1 represents full color. Default = .50.}
}
\value{
nice X-head plot with aggregated marker movements
}
\description{
Plots symbolic standardized head with markers plotted on predefined areas. 
Be sure to use standardized data, done by the functions \code{\link{bu2mm}}
and \code{\link{scaleFaceData}}.
}
\examples{
\dontrun{plotMarkerXHead()}

}
\author{
Rainer Alexandrowicz, Axel Zinkernagel
  \email{rainer.alexandrowicz@aau.at, zinkernagel@uni-landau.de}
}

