% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotOutliers.R
\name{plotOutliers}
\alias{plotOutliers}
\title{Plot potential outliers for specific frames}
\usage{
plotOutliers(data, colNameFrames, colNameData, colNameCond = "",
  SDFactor = 3, title = "", verbose = FALSE)
}
\arguments{
\item{data}{Data-frame containing the variables labeled in colNameFrames,
colNameData, and colNameCond.}

\item{colNameFrames}{Character vector containing the column name of the 
frames column of the data data-frame}

\item{colNameData}{Character vector containing the column name of the 
data which should be analyzed for outliers (e.g., AU_01_L_x)}

\item{colNameCond}{Character vector containing the column name of the
stimuli presented at each frame. Default is no colNameCond.}

\item{SDFactor}{Numeric vector of length 1. Factor the SD is multiplied with.
Default is 3.}

\item{title}{Default is no title.}

\item{verbose}{Logical value. If verbose is TRUE, the plots for each outlier
will be presented separately.}
}
\value{
See vignette for example plots.
}
\description{
The function computes the difference of the marker coordinate (x, y, or z) of 
frame \emph{t} with frame \emph{t + 1}. An outlier is identified by hitting 
a cut off definded by a multiplicative of the standard deviation (SD). The SD
is computed for the whole difference vector. The default cut off value is 
3 SD. The function plots for each potential outlier the movemnent of the 
centered marker coordinate (x, y, or z), the deviation of the movement, the upper 
and lower cut off values and marks the frame with outlier. Use the function
ahead of code{\link{centerCond}} otherwise the centering may be identified
as an outlier.
}
\examples{
\dontrun{
# fix me: meaningful example
}   
}
\author{
Axel Zinkernagel \email{zinkernagel@uni-landau.de}, Rainer Alexandrowicz \email{rainer.alexandrowicz@aau.at}
}
