% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotXhead.R
\name{plotXhead}
\alias{plotXhead}
\title{Plot markers on symbolic head}
\usage{
plotXhead(data, dataPos, title = "", overplot = FALSE, color = "black",
  alpha = 0.5, plotDataPos = FALSE)
}
\arguments{
\item{data}{Data-frame with a row sequence of *standardized* x/y
coordinates for each marker. Please use the function \code{\link{face2stdFace}} 
before plotting the data. Names of the data columns 
must and in '_x' for the x-axis and in '_y' for the y-axis.}

\item{dataPos}{Named list of x/y starting position coordinates for each marker which
should be plotted on the standardized face (e.g., dataPos = 
c(AU_01_L = c(-.3,.7), AU_01_R = c(.3,.7)), ...). Axel: fix me: This vector can also be generated by using 
the offsetTable generated by the function \code{\link{centerCond}}. Due to the tracking
procedure in blender, the center of the coordinate system is the nose tip 
at x = 0, y = 0.}

\item{title}{Default is no title.}

\item{overplot}{If TRUE, only the marker plots, but no base graph is plotted.
Default is FALSE.}

\item{color}{Character vector of lenght one to define the color for plotting
the Marker. Default is 'black'.}

\item{alpha}{Alpha value / density for marker plots. 0 represents full 
transparency, whereas 1 represents full color. Default = .50.}

\item{plotDataPos}{If TRUE, for debugging puposes the start position coordinates of dataPos are plotted.
Default is FALSE.}
}
\value{
nice X-head plot with marker movements
}
\description{
Plots symbolic standardized head with markers plotted on predefined areas. 
Be sure to use standardized data, done by the function \code{\link{face2stdFace}}.
}
\examples{
\dontrun{plotXhead()}

}
\author{
Rainer Alexandrowicz \email{rainer.alexandrowicz@aau.at}, 
Axel Zinkernagel \email{zinkernagel@uni-landau.de}
}

