% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotXYmmpf.R
\name{plotXYmmpf}
\alias{plotXYmmpf}
\title{Plot x-, y-axis marker movement per frame}
\usage{
plotXYmmpf(colFrames, colX, colY, colCond = "", center = TRUE,
  color = c("red", "orange"), title = "", overplot = FALSE)
}
\arguments{
\item{colFrames}{Numeric vector containing the frames for which x/y-values
should be plotted}

\item{colX}{Numeric vector with x values (same length as frames vector)}

\item{colY}{Numeric vector with y values (same length as frames vector)}

\item{colCond}{Character vector containing the stimuli presented at each frame (same length as frames vector). Stimulus labels must be unique. Optional. Default is NA.}

\item{center}{If TRUE, center x/y values by setting the x/y values to 0 at
the first (appropriate, e.g. non NA frame) frame. Default = TRUE}

\item{color}{Character vector of two values to define color for x/y values of
the Marker. Default is c('red','orange').}

\item{title}{Default is no title.}

\item{overplot}{If TRUE, only the x/y lines, but no base graph is plotted.
Default is FALSE.}
}
\value{
See vignette for example plots.
}
\description{
Plots separate lines for the x,y marker values (y-axis) per frame (x-axis).
}
\examples{
# Select data of a subject at relevant frame sections
data_Subj2Cen <- subset(dataStdFCen, subset = ((dataSmm$subject == 2) & 
                       (dataSmm$Frame >= 690)& (dataSmm$Frame <= 1610)))
plotXYmmpf(colFrames = data_Subj2Cen$Frame, colX = data_Subj2Cen$BR4_x, 
           colY = data_Subj2Cen$BR4_y, colCond = data_Subj2Cen$Stimulustype, 
           center = FALSE, title = "Subject 2, BR4")
  
}
\author{
Axel Zinkernagel \email{zinkernagel@uni-landau.de}, Rainer Alexandrowicz \email{rainer.alexandrowicz@aau.at}
}
