% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_Function_ScaleBlenderData.R
\name{scaleBlenderData}
\alias{scaleBlenderData}
\title{Scale blender data}
\usage{
scaleBlenderData(data, colNames, colNameSubj, scaleFactor, rwMeasure = 8,
  verbose = FALSE)
}
\arguments{
\item{data}{Data frame containig the columns which should be rescaled.
Remaining colums will be returned untouched. The data frame should contain 
many subjects, e.g. as generated by the function ConcatData().}

\item{colNames}{Character vector with names of the columns of the data 
data-frame which should be rescaled.}

\item{colNameSubj}{Character vector with a single value containig the name 
of the subject column of the data data-frame.}

\item{scaleFactor}{Numeric vector containing the values of the real world
object measured in blender for each subject in the data data-frame. For 
example, this can be the diameter of a glue dot measured in blender. 
The scaleFactor vector must contain a scale value for each subject in the 
data data-frame. The sorting of subjects in data data-frame and 
scaleFactor vector must be the same.}

\item{rwMeasure}{Optional. Real world measure of the object diameter used to
rescale the blender data. The default is the diameter of a glue dot of 
8 millimeter.}

\item{verbose}{If TRUE, the function prints verbose output. Otherwise not.}
}
\value{
Rescaled input data frame.
}
\description{
Scales columns of a data frame according to a scale factor. For example, the
scale factor can be the diameter of a glue dot measured in blender. The
columns are rescaled accordingly to represent Millimeter instead of blender
units as measurement unit.
}
\examples{
\dontrun{scaleBlenderData()}

}
\author{
Axel Zinkernagel \email{zinkernagel@uni-landau.de}
}

