% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_Function_ScaleFaceData.R
\name{scaleFaceData}
\alias{scaleFaceData}
\title{Scale head proportions}
\usage{
scaleFaceData(data, colNames, colNameSubj, pupilDist, leftPMDist, scaleFactor,
  mouthcornerDist = NA, rightPMDist = NA, rwMeasure = 8,
  verbose = FALSE)
}
\arguments{
\item{data}{Data frame containig the columns which should be rescaled.
Remaining colums will be returned untouched (e.g., z-values because we have
no information to rescale them).}

\item{colNames}{Character vector with names of the columns which should be
rescaled. To differ between x-, and y-axis scaling, the variable names containing horizontal movement (left-right) should end with '_x',
whereas variable names containing vertical movement (up-down) should end with '_y'.}

\item{colNameSubj}{Character vector with a single value containig the name 
of the subject column of the data data-frame.}

\item{pupilDist}{Numeric vector of measurement of the distance from left to right pupil for each subject,
measured in blender units.}

\item{leftPMDist}{Numeric vector of measuremet of the distance from left pupil to left
mouthcorner for each subject, measured in blender units.}

\item{scaleFactor}{Numeric vector containing the values of the real world
object measured in blender for each subject in the data data-frame. For 
example, this can be the diameter of a glue dot measured in blender. 
The scaleFactor vector must contain a scale value for each subject in the 
data data-frame. The sorting of subjects in data data-frame and 
scaleFactor vector must be the same. This factor is used to rescale the distances which are 
measured in blender units.}

\item{mouthcornerDist}{Numeric vector of measurement of the distance from left to right
mouthcorner for each subject measured in blender units.}

\item{rightPMDist}{Numeric vector of measurement of the distance from right pupil to right
mouthcorner for each subject, measured in blender units.}

\item{rwMeasure}{Optional. Real world measure of the object diameter used to
rescale. Default is the diameter of a glue dot of 8 millimeter.}

\item{verbose}{If TRUE, the function prints verbose output. Otherwise not.}
}
\value{
Data frame with scaled columns.
}
\description{
The heads of the participants are of different sizes. This function rescales
the x/y values according to a standardized face. For resizing
the x-axis, the distance between the pupils of the subject is used. For resizing the y-axis,
the distance between the left eye and the left mouthcorner is used. 
Optionally, the distance between the mouthcorners for the x-axis, and the distance between
the right eye and the right mouthcorner for the y-axis may additionally be provided.
If provided, a mean of both measures is computed.
}
\examples{
\dontrun{scaleFaceData()}

}
\author{
Axel Zinkernagel, Rainer Alexandrowicz
  \email{zinkernagel@uni-landau.de, rainer.alexandrowicz@aau.at}
}

