% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/angleDist.R
\name{angleDist}
\alias{angleDist}
\title{Computes distance and angle from marker columns}
\usage{
angleDist(data, colNames, colNameSubj, rndDig = 2, verbose = FALSE)
}
\arguments{
\item{data}{Data frame containig the data for a subset, defined in colNames, 
for which the angle and the distance should be computed for.}

\item{colNames}{Character vector naming the two columns with the x/y-Values
for which the angle and the distance should be computed.}

\item{colNameSubj}{Character vector with a single value containig the name 
of the subject column of the data data-frame.}

\item{rndDig}{Number of round digits. Default is 2.}

\item{verbose}{If TRUE, the function prints verbose output. Otherwise not. 
Default is FALSE.}
}
\value{
Numeric matrix with median distance and angle.
}
\description{
Computes the distance and the angle for a (mean, median) x/y marker 
point. The distance and the angle is computed from x = 0, y = 0 to the 
x/y - values given. Therefore, to get correct values for distance and angle
use the function \code{\link{centerCond}} before, to set the x/y-start 
values for each condition to x = 0, y = 0.
}
\examples{
\dontrun{angleDistance(x,y)}

}
\author{
Axel Zinkernagel \email{zinkernagel@uni-landau.de}
}

