% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/angleDist.R
\name{angleDist}
\alias{angleDist}
\title{Computes distance and angle from marker columns}
\usage{
angleDist(data, colNames, colNameSubj, rndDig = 2, verbose = FALSE)
}
\arguments{
\item{data}{Data frame containig the data for a subset, defined in colNames, 
for which the angle and the distance should be computed for.}

\item{colNames}{Character vector naming the two columns with the x/y-Values
for which the angle and the distance should be computed.}

\item{colNameSubj}{Character vector with a single value containig the name 
of the subject column of the data data-frame.}

\item{rndDig}{Number of round digits. Default is 2.}

\item{verbose}{If TRUE, the function prints verbose output. Otherwise not. 
Default is FALSE.}
}
\value{
Numeric matrix with median distance and angle.
}
\description{
Computes the distance and the angle for a (mean, median) x/y marker 
point. The distance and the angle is computed from x = 0, y = 0 to the 
x/y - values given. Therefore, to get correct values for distance and angle
use the function \code{\link{centerCond}} before, to set the x/y-start 
values for each condition to x = 0, y = 0.
}
\examples{
colNames <- c("AU_01_L_x", "AU_01_L_y", "AU_01_R_x", "AU_01_R_y", 
              "AU_02_L_x", "AU_02_L_y", "AU_02_R_x", "AU_02_R_y", 
              "AU_06_L_x", "AU_06_L_y", "AU_06_R_x", "AU_06_R_y", 
              "AU_08_x", "AU_08_y", 
              "AU_09_L_x", "AU_09_L_y", "AU_09_R_x", "AU_09_R_y", 
              "AU_10_L_x", "AU_10_L_y", "AU_10_R_x", "AU_10_R_y",  
              "AU_11_L_x", "AU_11_L_y", "AU_11_R_x", "AU_11_R_y",
              "AU_12_L_x", "AU_12_L_y", "AU_12_R_x", "AU_12_R_y", 
              "AU_16_x", "AU_16_y")

# Data preparation
data_Subj_happy <- subset(dataSmm, subset = dataSmm["Stimulustype"] == "posed_happy",
                          select = c("subject",colNames))
angleDist(data_Subj_happy, colNames = c("AU_12_L_x", "AU_12_L_y"),
          colNameSubj = "subject", rndDig = 3)

}
\author{
Axel Zinkernagel \email{zinkernagel@uni-landau.de}
}
