% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_Function_AngleDistance.R
\name{AngleDistance}
\alias{AngleDistance}
\title{Computes distance and angle from marker columns}
\usage{
AngleDistance(x, y, verbose = FALSE)
}
\arguments{
\item{x}{Numeric vector of x values of x marker movement, beginnig at 0}

\item{y}{Numeric vector of y values of y marker movement, beginnig at 0}

\item{verbose}{If TRUE, the function prints verbose output. Otherwise not.}
}
\value{
Numeric vector with median distance and angle.
}
\description{
Computes the distance and the angle for x/y marker movement. Data is supposed
to start at x = 0, y =0, use function CenterCond before.
}
\examples{
\dontrun{AngleDistance(x,y)}

}
\author{
Axel Zinkernagel \email{zinkernagel@uni-landau.de}
}

