% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centerCond.R
\name{centerCond}
\alias{centerCond}
\title{Center marker coordinates}
\usage{
centerCond(data, colNames, colNameSubj, colNameFrames, colNameCond,
  maxCPUcores = 4, verbose = FALSE)
}
\arguments{
\item{data}{Data frame containig the columns which should be centered and 
additional columns defining subject number, stimulus conditions, frames, 
etc. Columns not to be centered will return unchanged.}

\item{colNames}{Character vector with names of the columns which should be
centered.}

\item{colNameSubj}{Character vector with a single value containig the name of
the subject column. It is assumed, that each experimental condition
occurres only once per subject}

\item{colNameFrames}{Character vector with a single value containig the name
of the frames column.}

\item{colNameCond}{Character vector with a single value containing the
experimental conditions column name of the data frame.}

\item{maxCPUcores}{Optional integer value. Per default, all available CPU 
cores up to 4 are used. If more or less cores should be used, set it via 
the parameter \code{maxCPUcores} (e.g., \code{maxCPUcores} = 12).}

\item{verbose}{If TRUE, the function prints verbose output. Otherwise not.}
}
\value{
Data frame with centered columns per subject and per stimulus condition.
}
\description{
Sets the x/y/z-values of markers in parameter \code{colNames} defined 
variable columns to 0 at the start of every condition defined in parameter 
\code{colNameCond}. This is done for every subject defined by column 
\code{colNameSubj}.
}
\examples{
colNames <- c("AU_01_L_x", "AU_01_L_y", "AU_01_R_x", "AU_01_R_y", 
              "AU_02_L_x", "AU_02_L_y", "AU_02_R_x", "AU_02_R_y", 
              "AU_06_L_x", "AU_06_L_y", "AU_06_R_x", "AU_06_R_y", 
              "AU_08_x", "AU_08_y", 
              "AU_09_L_x", "AU_09_L_y", "AU_09_R_x", "AU_09_R_y", 
              "AU_10_L_x", "AU_10_L_y", "AU_10_R_x", "AU_10_R_y",  
              "AU_12_L_x", "AU_12_L_y", "AU_12_R_x", "AU_12_R_y", 
              "AU_16_x", "AU_16_y")
              
# To not overwrite existing data, use a new data frame 
# (dataStdFCen means data of standaradized faces, centered)

dataStdFCen <- centerCond(dataStdF, colNames = colNames, 
                          colNameSubj = "subject", colNameFrames = "Frame", 
                          colNameCond = "Stimulustype", maxCPUcores = 1,
                          verbose = TRUE)

}
\author{
Axel Zinkernagel \email{zinkernagel@uni-landau.de}
}
