% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concatBlenderFiles.R
\name{concatBlenderFiles}
\alias{concatBlenderFiles}
\title{Concatenate Blender data files}
\usage{
concatBlenderFiles(
  dataFileNames,
  inputDirectory,
  colNameSubj = "",
  verbose = FALSE
)
}
\arguments{
\item{dataFileNames}{Character vector containing filenames to be concatenated. 
Possible file types are csv and R data files (file extensions: *.csv, *.rda, 
*.Rdata).}

\item{inputDirectory}{Character vector of length one containing the path to the 
directory of data files (e.g., on Windows: 'C:/Data/Blenderdata/', on 
Unix-type systems: '/home/user/Data/Blenderdata/').}

\item{colNameSubj}{Character vector of length one, default is empty (""). 
If empty, the subject number is taken from the last number in the file names 
of the single data files, which should be concatenated (in front of the file 
extension). For example, from the filename 'RawData_Subj_39.RData' the subject number 
 39 is taken. If the subject number is contained the single data files, the
colNameSubj parameter must contain the column name of the subject column of
the data files (e.g., \code{colNameSubj = "subject"}).}

\item{verbose}{Logical value. If TRUE, the function provides verbose console output.}
}
\value{
Returns data frame and saves Rdata file of concatenated input files.
}
\description{
Concatenate several csv or Rdata files into one large 'master'-file. Either csv
or Rdata files can be concatenated, however, not mixed cvs/Rdata files. The data 
files must all reside in the same directory. The files need not to have the 
same number of colums, e.g., it is possible to concatenate files with a different
number of markers tracked. The total number of columns will be
determined during runtime of this function. Before concatenating the files, 
a feedback of the columns for the final data set is given and asked for it's
correctness.
}
\examples{
\dontrun{
inputdir <- paste(system.file(package = "blenderFace"),"/extdata/",sep="")
outputdir <- paste(system.file(package = "blenderFace"),"/data/",sep="")
filenames <- c("Subject_01.csv","Subject_02.csv")

concatBlenderFiles(dataFileNames = filenames, inputDirectory = inputdir, 
colNameSubj = "", verbose = TRUE)
}

}
\author{
Axel Zinkernagel \email{zinkernagel@uni-wuppertal.de},
Rainer Alexandrowicz \email{rainer.alexandrowicz@aau.at}
}
