% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_Function_ConcatData.R
\name{concatData}
\alias{concatData}
\title{Concatenate data files}
\usage{
concatData(dataFileNames, inputDirectory, subjectColumn = FALSE,
  outputFilename, outputDirectory = "", verbose = FALSE)
}
\arguments{
\item{dataFileNames}{character vector of the filenames to be concatenated. Possible
filetypes are csv and R data files (File endings: *.csv, *.rda, *.Rdata).}

\item{inputDirectory}{Character variable containing the path to the input
files (e.g., on Windows: 'C:/Data/Blenderdata/', on Linux:
'/home/user/Data/Blenderdata/). Should end with a backslash.}

\item{subjectColumn}{Logical value. Default is FALSE. Do the single data files contain a column with
the subject number?  If TRUE, the subject number is taken from the single files. If FALSE, the subject number has to be part of the
filename. It should be the last number before the filetype (e.g., '.RData' or '.csv') divided by an underscore from the rest of the 
filename. For example, from the filename 'RawData_Vpn_39.RData' the subject number 39 is generated.}

\item{outputFilename}{Name of the output file}

\item{outputDirectory}{Optional. Path to where the output file should be
saved (e.g., on Windows: 'C:/Data/Blenderdata/output/', on Linux:
'/home/user/Data/Blenderdata/output/). If empty, path of inputDirectory is
beeing used. Should end with a backslash (or slash, respectively).}

\item{verbose}{If TRUE, the function prints verbose output.}
}
\value{
Returns data frame and saves Rdata file of concatenated input files.
}
\description{
Concatenate several csv/Rdata files into one large 'master'-file. Either csv
or Rdata files can be concatenated, but not mixed cvs/Rdata files. Data files must not
have the same number of colums, e.g., it is possible to concatenate files with a different number of markers tracked. The exhaustive number of columns will be
determined by this function.
}
\examples{
\dontrun{concatData()}

}
\author{
Axel Zinkernagel \email{zinkernagel@uni-landau.de}
}

