% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotIndmm.R
\name{plotIndmm}
\alias{plotIndmm}
\title{Plots median movement of a marker per subject}
\usage{
plotIndmm(
  data,
  colNames,
  colNameSubj,
  title = "",
  xlim = NA,
  ylim = NA,
  verbose = FALSE
)
}
\arguments{
\item{data}{Data frame containig the data columns to be plotted (among other
columns).}

\item{colNames}{Character vector naming the two columns for with the x/y-Values
should be plotted.}

\item{colNameSubj}{Character vector with a single value containig the name 
of the subject column of the data data-frame.}

\item{title}{Character string containing the title. Default is no title.}

\item{xlim}{Numeric vector of two elements defining the x-axis range of the 
plot. Default is empty (NA), to compute the min and max values to achieve 
a quadratic plot.}

\item{ylim}{Numeric vector of two elements defining the y-axis range of the 
plot. Default is empty (NA), to compute the min and max values to achieve 
a quadratic plot.}

\item{verbose}{Logical value. If TRUE, the function provides verbose 
console output.}
}
\value{
See vignette for example plots.
}
\description{
Plots median movement of a marker per subject and labels it with subject 
number. This function may be used to identify individual outliers. To achieve 
meaninful plots, use the function \code{\link{centerCond}} first.
}
\examples{
colNames <- c("A7_x",  "A7_y",  "A8_x",  "A8_y",  
              "BL2_x", "BL2_y", "BL4_x", "BL4_y",  
              "BL5_x", "BL5_y", "BL7_x", "BL7_y",        
              "BR2_x", "BR2_y", "BR4_x", "BR4_y",  
              "BR5_x", "BR5_y", "BR7_x", "BR7_y",  
              "CL4_x", "CL4_y", "CL7_x", "CL7_y",        
              "CR4_x", "CR4_y", "CR7_x", "CR7_y",  
              "DL2_x", "DL2_y", "DR2_x", "DR2_y")
              
# Select data for plotting (selecting stimulus type and omit z-axis)
data_Subj_happy <- subset(dataStdFCen, subset = (dataStdFCen$Stimulustype == "posed_happy"), 
                          select = c("subject",colNames))

plotIndmm(data = data_Subj_happy, colNames = c("CL7_x", "CL7_y"), 
          colNameSubj = "subject", title = "Posed Happy CL7")
}
\author{
Axel Zinkernagel \email{zinkernagel@uni-landau.de}
}
