% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_Function_PlotXHead.R
\name{PlotMarkerXHead}
\alias{PlotMarkerXHead}
\title{Plot markers on symbolic head}
\usage{
PlotMarkerXHead(data, dataPos, dataScale, center = TRUE, title = "",
  overplot = FALSE, color = "black", alpha = 0.99, rwMeasure = 8)
}
\arguments{
\item{data}{matrix or data-frame with a row sequence of *standardized* x/y coordinates for each marker. Please use function "ScaleFaceData" to standardize data.}

\item{dataPos}{numeric vector with x/y starting coordinates each marker should be plotted on the face. Should have equal lenght to the number of columns of data. Default is FIX ME}

\item{dataScale}{numeric vector of four values containing measured left pupil - right pupil distance, left mouthcorner - right mouthcorner distance, left pupil - left mouthcorner distance, and right pupil - right mouthconrer distance.}

\item{center}{if TRUE, center x/y values by setting the x/y values to 0 at the first (appropriate, e.g. non NA frame) frame. Default = TRUE}

\item{title}{Default is no title.}

\item{overplot}{If TRUE, only the marker plots, but no base graph is plotted. Default is FALSE.}

\item{color}{character vector of lenght one to define the color for plotting the Marker. Default is "black".}

\item{alpha}{alpha value / density for marker plots. Default = .99}

\item{rwMeasure}{Optional. Real world measure of the object diameter used to rescale. Default is the diameter of a glue dot of 8 millimeter.}
}
\value{
nice "Quadratschädel" (square head) Plot
}
\description{
Plots symbolic standardized head with markers plotted on predefined areas.
}
\examples{
PlotMarkerXHead()

}
\author{
Rainer Alexandrowicz, Axel Zinkernagel \email{rainer.alexandrowicz@aau.at, zinkernagel@uni-landau.de}
}

