% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMmpCond.R
\name{plotMmpCond}
\alias{plotMmpCond}
\title{Plot marker and quartile ellipse per condition}
\usage{
plotMmpCond(
  data,
  colNames,
  colNameCond,
  title = "",
  xlim = NA,
  ylim = NA,
  verbose = FALSE
)
}
\arguments{
\item{data}{Data frame containig the data to be plotted.}

\item{colNames}{Character vector naming the two columns with the x/y-Values
to be plotted.}

\item{colNameCond}{Character vector with a single value containig the name 
of the stimulus type column of the data data-frame.}

\item{title}{Default is no title.}

\item{xlim}{Numeric vector of two elements defining the x-axis range of the 
plot. Per default the scaling is done automatically.}

\item{ylim}{Numeric vector of two elements defining the y-axis range of the 
plot. Per default the scaling is done automatically.}

\item{verbose}{Logical value. If TRUE, the function provides verbose 
console output.}
}
\value{
See vignette for example plots.
}
\description{
Plots the median and quartile ellipse per experimental condition
}
\examples{
plotMmpCond(data = dataStdFCen, colNames = c("CL7_x", "CL7_y"), 
            colNameCond = "Stimulustype", title = "CL7")
   
}
\author{
Axel Zinkernagel \email{zinkernagel@uni-landau.de}
}
