% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotXhead.R
\name{plotXhead}
\alias{plotXhead}
\title{Plot markers on symbolic head}
\usage{
plotXhead(data, dataPos, title = "", overplot = FALSE, color = "black",
  alpha = 0.5, plotDataPos = FALSE)
}
\arguments{
\item{data}{Data-frame with a row sequence of standardized and centered x/y
coordinates for each marker. Please use the functions 
\code{\link{face2stdFace}} and \code{\link{centerCond}} 
before plotting the data. Names of the data columns 
must end in '_x' for the x-axis and in '_y' for the y-axis.}

\item{dataPos}{Named list of x/y starting position coordinates for each marker which
should be plotted on the standardized face (e.g., dataPos = 
c(AU_01_L = c(-.3,.7), AU_01_R = c(.3,.7)), ...). Due to the tracking
procedure in blender, the center of the coordinate system is the nose tip 
at x = 0, y = 0.}

\item{title}{Default is no title.}

\item{overplot}{If TRUE, only the marker plots, but no base graph is plotted.
Default is FALSE.}

\item{color}{Character vector of length one to define the color for plotting
the Marker. Default is 'black'.}

\item{alpha}{Alpha value / density for marker plots. 0 represents full 
transparency, whereas 1 represents full color. Default = .50. This may be 
  adapted for plotting many participants}

\item{plotDataPos}{If TRUE, for debugging puposes, the start position 
coordinates of dataPos are plotted. Default is FALSE.}
}
\value{
See vignette for example plots.
}
\description{
Plots a symbolic, standardized head with markers plotted at predefined 
starting points. To get meaningful plots, functions 
\code{\link{face2stdFace}} and \code{\link{centerCond}} first.
}
\examples{
colNames <- c("AU_01_L_x", "AU_01_L_y", "AU_01_R_x", "AU_01_R_y", 
              "AU_02_L_x", "AU_02_L_y", "AU_02_R_x", "AU_02_R_y", 
              "AU_06_L_x", "AU_06_L_y", "AU_06_R_x", "AU_06_R_y", 
              "AU_08_x", "AU_08_y", 
              "AU_09_L_x", "AU_09_L_y", "AU_09_R_x", "AU_09_R_y", 
              "AU_10_L_x", "AU_10_L_y", "AU_10_R_x", "AU_10_R_y",
              "AU_11_L_x", "AU_11_L_y", "AU_11_R_x", "AU_11_R_y",  
              "AU_12_L_x", "AU_12_L_y", "AU_12_R_x", "AU_12_R_y", 
              "AU_16_x", "AU_16_y")

# Select data for plotting (selecting stimulus type and omit z-axis)
data_Subj_happy <- subset(dataStdFCen, subset = (dataStdFCen$Stimulustype == "posed_happy"), select = colNames)

# Define the positions for the markers for the standardized face of x (-1,1) and y (-1,1) size as named list
dataPos <- list(AU_01_L = c(-.3,.7), AU_01_R = c(.3,.7), AU_02_L = c(-.7,.7), AU_02_R = c(.7,.7), 
                AU_06_L = c(-.5,.2), AU_06_R = c(.5,.2), AU_08 = c(0,-.55), 
                AU_09_L = c(-.2,.2), AU_09_R = c(.2,.2), AU_10_L = c(-.2,-.6), AU_10_R = c(.2,-.6), 
                AU_11_L = c(-.2,-.1), AU_11_R = c(.2,-.1), AU_12_L = c(-.3,-.7), AU_12_R = c(.3,-.7), 
                AU_16 = c(0,-.8)) 

# For debugging purposes the marker names and start positions may also be plotted
plotXhead(data = data_Subj_happy[-1], dataPos = dataPos, title = "All Subjects, happy", plotDataPos = TRUE)

}
\author{
Rainer Alexandrowicz \email{rainer.alexandrowicz@aau.at}, 
Axel Zinkernagel \email{zinkernagel@uni-landau.de}
}

