% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotXYmmpf.R
\name{plotXYmmpf}
\alias{plotXYmmpf}
\title{Plot x-, y-axis marker movement per frame}
\usage{
plotXYmmpf(frames, xMarker, yMarker, stimF = "", center = TRUE,
  color = c("red", "orange"), title = "", overplot = FALSE)
}
\arguments{
\item{frames}{numeric vector containing the frames for which x/y-values
should be plotted}

\item{xMarker}{numeric vector with x values (same length as frames vector)}

\item{yMarker}{numeric vector with y values (same length as frames vector)}

\item{stimF}{character vector containing the stimuli presented at each frame (same length as frames vector). Stimulus labels must be unique. Optional. Default is NA.}

\item{center}{if TRUE, center x/y values by setting the x/y values to 0 at
the first (appropriate, e.g. non NA frame) frame. Default = TRUE}

\item{color}{character vector of two values to define color for x/y values of
the Marker. Default is c('red','orange').}

\item{title}{Default is no title.}

\item{overplot}{If TRUE, only the x/y lines, but no base graph is plotted.
Default is FALSE.}
}
\value{
See vignette for example plots.
}
\description{
Plots separate lines for the x,y marker values (y-axis) per frame (x-axis).
}
\examples{
\dontrun{plotXYmmpf()}
  
}
\author{
Axel Zinkernagel \email{zinkernagel@uni-landau.de}, Rainer Alexandrowicz \email{rainer.alexandrowicz@aau.at}
}

