% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_Function_ScaleBlenderData.R
\name{ScaleBlenderData}
\alias{ScaleBlenderData}
\title{Scale blender data}
\usage{
ScaleBlenderData(data, colNames, scaleFactor, rwMeasure = 8,
  verbose = FALSE)
}
\arguments{
\item{data}{Data frame containig the columns which should be rescaled.
Remaining colums will be returned untouched.}

\item{colNames}{Character vector with names of the columns which should be
rescaled.}

\item{scaleFactor}{Numeric variable containing the value of the real world
object measured in blender. For example, this can be the diameter of a glue
dot measured in blender (default = 8 mm). Therefore, scaling should be done
person by person.}

\item{rwMeasure}{Optional. Real world measure of the object diamoter used to
rescale. Default is the diameter of a glue dot of 8 millimeter.}

\item{verbose}{If TRUE, the function prints verbose output. Otherwise not.}
}
\value{
Data frame with scaled columns.
}
\description{
Scales columns of a data frame according to a scale factor. For example, the
scale factor can be the diameter of a glue dot measured in blender. The
columns are rescaled accordingly to represent Millimeter instead of blender
units as measurement unit.
}
\examples{
\dontrun{ScaleBlenderData()}

}
\author{
Axel Zinkernagel \email{zinkernagel@uni-landau.de}
}

